chore: merge utils and utils.table
