--- src/utils.table.ts
@@ -1,199 +1,0 @@
-import process from 'node:process';
-import Table from 'cli-table3';
-import stringWidth from 'string-width';
-
-/**
- * Table row data type supporting strings, numbers, and formatted cell objects
- */
-type TableRow = (string | number | { content: string; hAlign?: 'left' | 'right' | 'center' })[];
-/**
- * Configuration options for creating responsive tables
- */
-type TableOptions = {
-	head: string[];
-	colAligns?: ('left' | 'right' | 'center')[];
-	style?: {
-		head?: string[];
-	};
-	dateFormatter?: (dateStr: string) => string;
-};
-
-/**
- * Responsive table class that adapts column widths based on terminal size
- * Automatically adjusts formatting and layout for different screen sizes
- */
-export class ResponsiveTable {
-	private head: string[];
-	private rows: TableRow[] = [];
-	private colAligns: ('left' | 'right' | 'center')[];
-	private style?: { head?: string[] };
-	private dateFormatter?: (dateStr: string) => string;
-
-	/**
-	 * Creates a new responsive table instance
-	 * @param options - Table configuration options
-	 */
-	constructor(options: TableOptions) {
-		this.head = options.head;
-		this.colAligns = options.colAligns ?? Array.from({ length: this.head.length }, () => 'left');
-		this.style = options.style;
-		this.dateFormatter = options.dateFormatter;
-	}
-
-	/**
-	 * Adds a row to the table
-	 * @param row - Row data to add
-	 */
-	push(row: TableRow): void {
-		this.rows.push(row);
-	}
-
-	/**
-	 * Renders the table as a formatted string
-	 * Automatically adjusts layout based on terminal width
-	 * @returns Formatted table string
-	 */
-	toString(): string {
-		const terminalWidth = process.stdout.columns || 120;
-
-		// Calculate actual content widths first (excluding separator rows)
-		const dataRows = this.rows.filter(row => !this.isSeparatorRow(row));
-		const allRows = [this.head.map(String), ...dataRows.map(row => row.map((cell) => {
-			if (typeof cell === 'object' && cell != null && 'content' in cell) {
-				return String(cell.content);
-			}
-			return String(cell ?? '');
-		}))];
-		const contentWidths = this.head.map((_, colIndex) => {
-			const maxLength = Math.max(
-				...allRows.map(row => stringWidth(String(row[colIndex] ?? ''))),
-			);
-			return maxLength;
-		});
-
-		// Calculate table overhead
-		const numColumns = this.head.length;
-		const tableOverhead = 3 * numColumns + 1; // borders and separators
-		const availableWidth = terminalWidth - tableOverhead;
-
-		// Always use content-based widths with generous padding for numeric columns
-		const columnWidths = contentWidths.map((width, index) => {
-			const align = this.colAligns[index];
-			// For numeric columns, ensure generous width to prevent truncation
-			if (align === 'right') {
-				return Math.max(width + 3, 11); // At least 11 chars for numbers, +3 padding
-			}
-			else if (index === 1) {
-				// Models column - can be longer
-				return Math.max(width + 2, 15);
-			}
-			return Math.max(width + 2, 10); // Other columns
-		});
-
-		// Check if this fits in the terminal
-		const totalRequiredWidth = columnWidths.reduce((sum, width) => sum + width, 0) + tableOverhead;
-
-		if (totalRequiredWidth > terminalWidth) {
-			// Apply responsive resizing and use compact date format if available
-			const scaleFactor = availableWidth / columnWidths.reduce((sum, width) => sum + width, 0);
-			const adjustedWidths = columnWidths.map((width, index) => {
-				const align = this.colAligns[index];
-				let adjustedWidth = Math.floor(width * scaleFactor);
-
-				// Apply minimum widths based on column type
-				if (align === 'right') {
-					adjustedWidth = Math.max(adjustedWidth, 10);
-				}
-				else if (index === 0) {
-					adjustedWidth = Math.max(adjustedWidth, 10);
-				}
-				else if (index === 1) {
-					adjustedWidth = Math.max(adjustedWidth, 12);
-				}
-				else {
-					adjustedWidth = Math.max(adjustedWidth, 8);
-				}
-
-				return adjustedWidth;
-			});
-
-			const table = new Table({
-				head: this.head,
-				style: this.style,
-				colAligns: this.colAligns,
-				colWidths: adjustedWidths,
-				wordWrap: true,
-				wrapOnWordBoundary: true,
-			});
-
-			// Add rows with special handling for separators and date formatting
-			for (const row of this.rows) {
-				if (this.isSeparatorRow(row)) {
-					// Skip separator rows - cli-table3 will handle borders automatically
-					continue;
-				}
-				else {
-					// Use compact date format for first column if dateFormatter available
-					const processedRow = row.map((cell, index) => {
-						if (index === 0 && this.dateFormatter != null && typeof cell === 'string' && this.isDateString(cell)) {
-							return this.dateFormatter(cell);
-						}
-						return cell;
-					});
-					table.push(processedRow);
-				}
-			}
-
-			return table.toString();
-		}
-		else {
-			// Use generous column widths with normal date format
-			const table = new Table({
-				head: this.head,
-				style: this.style,
-				colAligns: this.colAligns,
-				colWidths: columnWidths,
-				wordWrap: true,
-				wrapOnWordBoundary: true,
-			});
-
-			// Add rows with special handling for separators
-			for (const row of this.rows) {
-				if (this.isSeparatorRow(row)) {
-					// Skip separator rows - cli-table3 will handle borders automatically
-					continue;
-				}
-				else {
-					table.push(row);
-				}
-			}
-
-			return table.toString();
-		}
-	}
-
-	/**
-	 * Checks if a row is a separator row (contains only empty cells or dashes)
-	 * @param row - Row to check
-	 * @returns True if the row is a separator
-	 */
-	private isSeparatorRow(row: TableRow): boolean {
-		// Check for both old-style separator rows (─) and new-style empty rows
-		return row.every((cell) => {
-			if (typeof cell === 'object' && cell != null && 'content' in cell) {
-				return cell.content === '' || /^─+$/.test(cell.content);
-			}
-			return typeof cell === 'string' && (cell === '' || /^─+$/.test(cell));
-		});
-	}
-
-	/**
-	 * Checks if a string matches the YYYY-MM-DD date format
-	 * @param text - String to check
-	 * @returns True if the string is a valid date format
-	 */
-	private isDateString(text: string): boolean {
-		// Check if string matches date format YYYY-MM-DD
-		return /^\d{4}-\d{2}-\d{2}$/.test(text);
-	}
-}
